Add the navigation in the persional page
I reordered the path of files as I am going to specify the pages' content.
For this commit I add the navigation between personal page and the pet
home page, in this task I faced a problem, that the buttons show in
the self header cannot be pressed, I determined a z index for them, and they
can successfully be pressed.